app/requirements.txt
flask==3.0.3
google-auth==2.29.0
google-api-python-client==2.125.0
